chore: refactor administration pages
