+++ ui/src/admin/containers/chronograf/ProvidersPage.js
