+++ ui/src/admin/containers/influxdb/AdminInfluxDBPage.tsx
@@ -28,8 +28,8 @@ import {
 import PageSpinner from 'src/shared/components/PageSpinner'
 import UsersTable from 'src/admin/components/UsersTable'
 import RolesTable from 'src/admin/components/RolesTable'
-import QueriesPage from 'src/admin/containers/QueriesPage'
-import DatabaseManagerPage from 'src/admin/containers/DatabaseManagerPage'
+import QueriesPage from './QueriesPage'
+import DatabaseManagerPage from './DatabaseManagerPage'
 import {Page} from 'src/reusable_ui'
 import SubSections from 'src/shared/components/SubSections'
 import {ErrorHandling} from 'src/shared/decorators/errors'
