+++ ui/src/admin/containers/influxdb/DatabaseManagerPage.js
