+++ ui/src/admin/containers/influxdb/QueriesPage.js
