+++ ui/src/admin/index.js
@@ -1,4 +1,4 @@
-import AdminInfluxDBPage from './containers/AdminInfluxDBPage'
+import AdminInfluxDBPage from './containers/influxdb/AdminInfluxDBPage'
 import AdminChronografPage from './containers/chronograf/AdminChronografPage'
 
 export {AdminChronografPage, AdminInfluxDBPage}
